Apply float precision alignment in quantize_i2_s
Replaced explicit double casts and implicit double literals with native single-precision equivalents in the quantize_i2_s loop. This prevents type promotion overhead and allows for cleaner SIMD instructions. Also recorded this mathematical optimization learning in the Hugo journal.

+++ src/ggml-bitnet-mad.cpp
@@ -57,7 +57,8 @@ size_t quantize_i2_s(const float * src, void * dst, int64_t nrow, int64_t n_per_
     memset(dst, 0, n * sizeof(uint8_t) / 4);
     uint8_t* i2_weight = (uint8_t*)dst;
 
-    double max = 0;
+    // Hugo: Use float to maintain precision parity with src, avoiding scalar promotion overhead for SIMD
+    float max = 0.0f;
 
     // Scale-Invariant Fused Quantization:
     // We compute max dynamically, check the threshold directly,
@@ -65,11 +66,11 @@ size_t quantize_i2_s(const float * src, void * dst, int64_t nrow, int64_t n_per_
     for (int i = 0; i < n / QK_I2_S; i++) {
         for (int j = 0; j < QK_I2_S; j++) {
             int src_idx = i * QK_I2_S + j;
-            double val = (double)src[src_idx];
-            max = fmax(max, (double)fabs(val));
+            float val = src[src_idx];
+            max = fmaxf(max, fabsf(val));
 
             uint8_t q_val;
-            if (fabs(val) < 1e-6) {
+            if (fabsf(val) < 1e-6f) {
                 q_val = 1;
             } else {
                 q_val = val > 0 ? 2 : 0;
@@ -95,7 +96,8 @@ size_t quantize_i2_s(const float * src, void * dst, int64_t nrow, int64_t n_per_
     uint8_t* out = (uint8_t*)dst;
     memset(out, 0, (size_t)(n / 4));
 
-    double max = 0;
+    // Hugo: Use float to maintain precision parity with src, avoiding scalar promotion overhead for SIMD
+    float max = 0.0f;
 
     int64_t nrow4 = nrow / 4;
     for (int64_t rg = 0; rg < nrow4; rg++) {
@@ -107,20 +109,20 @@ size_t quantize_i2_s(const float * src, void * dst, int64_t nrow, int64_t n_per_
         int64_t base = rg * n_per_row;
 
         for (int64_t col = 0; col < n_per_row; col++) {
-            double v0 = (double)src[r0 * n_per_row + col];
-            double v1 = (double)src[r1 * n_per_row + col];
-            double v2 = (double)src[r2 * n_per_row + col];
-            double v3 = (double)src[r3 * n_per_row + col];
+            float v0 = src[r0 * n_per_row + col];
+            float v1 = src[r1 * n_per_row + col];
+            float v2 = src[r2 * n_per_row + col];
+            float v3 = src[r3 * n_per_row + col];
 
-            max = fmax(max, (double)fabs(v0));
-            max = fmax(max, (double)fabs(v1));
-            max = fmax(max, (double)fabs(v2));
-            max = fmax(max, (double)fabs(v3));
+            max = fmaxf(max, fabsf(v0));
+            max = fmaxf(max, fabsf(v1));
+            max = fmaxf(max, fabsf(v2));
+            max = fmaxf(max, fabsf(v3));
 
-            uint8_t q0 = fabs(v0) < 1e-6 ? 1 : (v0 > 0 ? 2 : 0);
-            uint8_t q1 = fabs(v1) < 1e-6 ? 1 : (v1 > 0 ? 2 : 0);
-            uint8_t q2 = fabs(v2) < 1e-6 ? 1 : (v2 > 0 ? 2 : 0);
-            uint8_t q3 = fabs(v3) < 1e-6 ? 1 : (v3 > 0 ? 2 : 0);
+            uint8_t q0 = fabsf(v0) < 1e-6f ? 1 : (v0 > 0 ? 2 : 0);
+            uint8_t q1 = fabsf(v1) < 1e-6f ? 1 : (v1 > 0 ? 2 : 0);
+            uint8_t q2 = fabsf(v2) < 1e-6f ? 1 : (v2 > 0 ? 2 : 0);
+            uint8_t q3 = fabsf(v3) < 1e-6f ? 1 : (v3 > 0 ? 2 : 0);
 
             uint8_t packed = (uint8_t)((q0 << 6) | (q1 << 4) | (q2 << 2) | (q3 << 0));
             out[base + col] = packed;
@@ -139,16 +141,17 @@ size_t quantize_i2_s(const float * src, void * dst, int64_t nrow, int64_t n_per_
     memset(dst, 0, n * sizeof(uint8_t) / 4);
     uint8_t* i2_weight = (uint8_t*)dst;
 
-    double max = 0;
+    // Hugo: Use float to maintain precision parity with src, avoiding scalar promotion overhead for SIMD
+    float max = 0.0f;
 
     for (int i = 0; i < n / QK_I2_S; i++) {
         for (int j = 0; j < QK_I2_S; j++) {
             int src_idx = i * QK_I2_S + j;
-            double val = (double)src[src_idx];
-            max = fmax(max, (double)fabs(val));
+            float val = src[src_idx];
+            max = fmaxf(max, fabsf(val));
 
             uint8_t q_val;
-            if (fabs(val) < 1e-6) {
+            if (fabsf(val) < 1e-6f) {
                 q_val = 1;
             } else {
                 q_val = val > 0 ? 2 : 0;
